Relocate classes with slf4j in the API used from both agent and ot
