+++ communication/src/main/java/datadog/communication/monitor/DDAgentStatsDConnection.java
@@ -9,8 +9,8 @@ import com.timgroup.statsd.NoOpStatsDClient;
 import com.timgroup.statsd.NonBlockingStatsDClientBuilder;
 import com.timgroup.statsd.StatsDClientErrorHandler;
 import datadog.trace.api.Config;
-import datadog.trace.api.IOLogger;
 import datadog.trace.api.Platform;
+import datadog.trace.relocate.api.IOLogger;
 import datadog.trace.util.AgentTaskScheduler;
 import datadog.trace.util.AgentThreadFactory;
 import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
