+++ dd-java-agent/agent-profiling/profiling-uploader/src/main/java/com/datadog/profiling/uploader/ProfileUploader.java
@@ -28,8 +28,8 @@ import datadog.common.container.ContainerInfo;
 import datadog.common.socket.NamedPipeSocketFactory;
 import datadog.common.socket.UnixDomainSocketFactory;
 import datadog.trace.api.Config;
-import datadog.trace.api.IOLogger;
 import datadog.trace.bootstrap.config.provider.ConfigProvider;
+import datadog.trace.relocate.api.IOLogger;
 import datadog.trace.util.AgentProxySelector;
 import datadog.trace.util.AgentThreadFactory;
 import java.io.File;
