+++ dd-java-agent/dd-java-agent.gradle
@@ -39,6 +39,8 @@ ext.generalShadowJarConfig = {
     relocate 'datadog.trace.common', 'datadog.trace.agent.common'
     relocate 'datadog.trace.core', 'datadog.trace.agent.core'
     relocate 'datadog.opentracing', 'datadog.trace.agent.ot'
+    // shadow things in internal API that has slf4j in the API and is accessed from core
+    relocate 'datadog.trace.relocate', 'datadog.trace.agent.relocate'
   }
 }
