+++ dd-trace-core/src/main/java/datadog/trace/common/writer/ddagent/DDAgentApi.java
@@ -9,8 +9,8 @@ import datadog.communication.ddagent.DDAgentFeaturesDiscovery;
 import datadog.communication.monitor.Counter;
 import datadog.communication.monitor.Monitoring;
 import datadog.communication.monitor.Recording;
-import datadog.trace.api.IOLogger;
 import datadog.trace.core.DDTraceCoreInfo;
+import datadog.trace.relocate.api.IOLogger;
 import java.io.IOException;
 import java.util.ArrayList;
 import java.util.HashMap;
