+++ internal-api/src/main/java/datadog/trace/relocate/api/RatelimitedLogger.java
@@ -1,4 +1,4 @@
-package datadog.trace.api;
+package datadog.trace.relocate.api;
 
 import datadog.trace.api.time.SystemTimeSource;
 import datadog.trace.api.time.TimeSource;
